Refactor: move security service components to application subpackge
