+++ src/main/java/com/palindrome/studit/global/config/security/application/TokenService.java
@@ -1,4 +1,4 @@
-package com.palindrome.studit.global.config.security;
+package com.palindrome.studit.global.config.security.application;
 
 import com.auth0.jwt.JWT;
 import com.auth0.jwt.algorithms.Algorithm;
