+++ src/main/java/com/palindrome/studit/global/config/security/config/OAuth2LoginSecurityConfig.java
@@ -1,4 +1,4 @@
-package com.palindrome.studit.global.config.security;
+package com.palindrome.studit.global.config.security.config;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
